Finito il parser per CML e spostato tutti i menu della main_window in CML

+++ src/ParserCML/hpp/CMLValue.hpp
@@ -18,7 +18,7 @@ struct CMLIdentifier {
 };
 
 struct CMLNumber {
-    double value;
+    float value;
 
     [[nodiscard]]
     std::string ToString() const;
